+++ src/serious_python/example/flet_example/lib/main.dart
@@ -23,7 +23,8 @@ void main() async {
 
   if (defaultTargetPlatform == TargetPlatform.windows) {
     // use TCP on Windows
-    var port = await getUnusedPort();
+    //var port = await getUnusedPort();
+    var port = 63777;
     pageUrl = "tcp://localhost:$port";
     environmentVariables["FLET_SERVER_PORT"] = port.toString();
   } else {
@@ -38,9 +39,11 @@ void main() async {
   runApp(FletApp(
     pageUrl: pageUrl,
     assetsDir: path.join(appDir, "assets"),
+    hideLoadingPage: true,
   ));
 }
 
+// Calling this method causes Windows firewall dialog to popup
 Future<int> getUnusedPort() {
   return ServerSocket.bind(InternetAddress.anyIPv4, 0).then((socket) {
     var port = socket.port;