+++ src/serious_python_android/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.6.0
+
+* `--verbose` flag - verbose output.
+* `--mobile` flag - (removes `.so`) from app dest archive.
+* `--web` flag for packaging for pyodide.
+* `--find-links` option for installing pip dependencies from alternative sources (indexes).
+* `--dep-mappings` for rewriting `flet` dependency to either `flet-embed` or `flet-pyodide`.
+* `--req-deps` for adding required dependencies like `flet-embed` or `flet-pyodide`.
+* `--platform` option for use with `sitecustomize.py` to tweak pip to pull platform-specific packages.
+* More structured regular output.
+* Bump deps versions.
+
 ## 0.5.1
 
 * Simplified Python initialization on Android.